+++ src/imio/smartweb/core/contents/pages/views.py
@@ -50,7 +50,15 @@ class PagesView(FolderView):
     def get_class(self, obj):
         section_type = obj.portal_type.split(".")[-1]
         return " ".join(
-            [section_type.lower(), obj.css_class or "", obj.bootstrap_css_class or ""]
+            filter(
+                None,
+                [
+                    section_type.lower(),
+                    obj.css_class or "",
+                    obj.bootstrap_css_class or "",
+                    self.background_style(obj) and "with-background" or "",
+                ],
+            )
         )
 
     def background_style(self, obj):